% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{d}
\alias{d}
\title{Date vector of Landsat scenes included in stN}
\usage{
data(d)
}
\description{
Vector containing dates of all Landsat scenes included in stN
}
\author{
Jennifer von Keyserlingk
}
\keyword{data}
