% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resIndSpatial.R
\name{resIndSpatial}
\alias{resIndSpatial}
\title{Runs function "resInd" on gridded time series and returns spatial output}
\usage{
resIndSpatial(x, dates, type = "irregular", sc = 1, order = 3,
  formula = response ~ (trend + harmon), h = 0.15, plevel = 0.05, dr,
  drd, s = 3, NV = NA, mc.cores = 1)
}
\arguments{
\item{x}{RasterBrick or rasterStack object}

\item{dates}{See \code{\link{resInd}}}

\item{type}{See \code{\link{resInd}}}

\item{sc}{See \code{\link{resInd}}}

\item{order}{See \code{\link{resInd}}}

\item{formula}{See \code{\link{resInd}}}

\item{h}{See \code{\link{resInd}}}

\item{plevel}{See \code{\link{resInd}}}

\item{dr}{See \code{\link{resInd}}}

\item{drd}{See \code{\link{resInd}}}

\item{s}{See \code{\link{resInd}}}

\item{NV}{See \code{\link{resInd}}}

\item{mc.cores}{Numeric. Number of cores used for the calcualtion. Default=1}
}
\value{
Large Raster stack with 14 layers:
'BPNumb': Total number of breakpoints in time series
'Initial NDVI': Mean of data during the first "s" years of the time series.
'Intercept': Linear model intercept
'DBP': Drought Break Point yes/no (1/0). Yes, if a breakpoint occurs in time interval set with parameter "dr".
'BpTime': Timing of breakpoint. Format: Decimal year. If more than one breakpoints occurs during time interval, the first one is selected.
'Timelag': Number of days between drought reference day set with parameter "drd" and breakpoint
'RecTrend': Slope of linear trend in segment succeeding "drought breakpoint".
'PreTrend': Slope of linear trend in segment preceeding "drought breakpoint".
'PreNDVI': Mean of data of "s" years before drought breakpoint, based on observed data values.
'MagObsA': Absolute difference of mean observed data "s" years before and after the "drought breakpoint".
'MagObsR': Relative difference of mean observed data "s" years before and after the "drought breakpoint".
'MagTrendA': Absolute difference between last value of trend prediction before and first value of trend prediction after drought breakpoint. Based on corrected trend for irregular data. Still, with irregular data, the height of the trend line does not seem robust.
'MagTrendR': Relative difference between last value of trend prediction before and first value of trend prediction after drought breakpoint. Based on corrected trend for irregular data. Still, with irregular data, the height of the trend line does not seem robust.
'AmpDiffR': Relative difference in mean amplitudes (based on sine and cosine terms of harmonic model) in segment before and after drought breakpoint.
}
\description{
Computes several change detection metrics based on a BFAST type change detection framework for breakpoints ocurring during a given time period specified by the user (i.e. years of a known drought event). In addition, the overall number of breakpoints in the time series, the overall mean of the data and the inital value of the data at the beginning of the time series is extracted. Applicable to to gridded time-series data.
}
\details{
The function \code{\link{resInd}} was designed to explore several change detection metrics that can be extracted from a BFAST type change detection approach in relation to drought. The extracted metrics are at the experimental stage. Not all metrics were found equally reliable: for an irregular time series the interecept of the linear trend line within segments (i.e. the height of the trend line after a breakpoint when plotted), was not stable. Therefore, the metrics relying on this information ("MagTrendA", "MagTrendR") are equally not robust. The slope of the trend line, however, was not affected by this, and appears to be a robust model parameter.
The function "resIndSpatial" requires the function "mc.calc" from the package "bfastSpatial" \url{https://github.com/loicdtx/bfastSpatial}, which you can install from github.
}
\examples{
#Load example raster data set (5x5 cells, including NA values)
data(RM25NA)

#Plot raster brick with NDVI scaling factor
#plot(stN*0.0001)

#Load date vector
data(dates)

##With package "bfastSpatial" you can extract information on acquisition date from typical Landsat file names (https://github.com/loicdtx/bfastSpatial)
#gs <- getSceneinfo(names(stN))
#d <- gs$date

\dontrun{
#Run function "resIndSpatial"
y <- resIndSpatial(stN, d, dr=c(2004.753,2008.751), drd=2004.753)

#Should return a raster stack containing 14 layers
}
}
\author{
Jennifer von Keyserlingk
}
