% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{stN}
\alias{stN}
\title{Formal class RasterBrick}
\usage{
data(stN)
}
\description{
5x5 RasterBrick containing 476 Landsta scenes between 1984-2011. Includes NA values. Located in the Randi forest area (southern Cyprus).
}
\examples{
data(stN)
library(raster)
plot(stN, 1:9)
}
\author{
NASA/ESA/Jennifer von Keyserlingk
}
\keyword{data}
